Refactoring: Extracting out creation of projections and mappers
