+++ Source/Integration/Adapters.cs
@@ -1,14 +1,6 @@
 using System.Reflection;
 using AutoMapper;
-using Cratis.Concepts;
-using Cratis.Events;
-using Cratis.Events.Projections;
-using Cratis.Events.Projections.Json;
-using Cratis.Extensions.Dolittle.EventStore;
 using Cratis.Types;
-using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
-using IEventStore = Cratis.Extensions.Dolittle.EventStore.IEventStore;
 
 namespace Aksio.Integration
 {
@@ -19,34 +11,26 @@ namespace Aksio.Integration
     {
         readonly ITypes _types;
         readonly IServiceProvider _serviceProvider;
-        readonly IEventStore _eventStore;
-        readonly IEventTypes _eventTypes;
-        readonly JsonProjectionSerializer _projectionSerializer;
-        readonly ILoggerFactory _loggerFactory;
+        readonly IAdapterProjectionFactory _adapterProjectionFactory;
+        readonly IAdapterMapperFactory _adapterMapperFactory;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="Adapters"/> class.
         /// </summary>
         /// <param name="types"><see cref="ITypes"/> for type discovery.</param>
         /// <param name="serviceProvider"><see cref="IServiceProvider"/> for getting instances from the IoC container.</param>
-        /// <param name="eventStore"><see cref="IEventStream"/> to use.</param>
-        /// <param name="eventTypes">The <see cref="IEventTypes"/> to use.</param>
-        /// <param name="projectionSerializer"><see cref="JsonProjectionSerializer"/> used for deserialization to Cratis Kernel ProjectionDefinition.</param>
-        /// <param name="loggerFactory"><see cref="ILoggerFactory"/> for creating loggers.</param>
+        /// <param name="adapterProjectionFactory"><see cref="IAdapterProjectionFactory"/> for creating projections for adapters.</param>
+        /// <param name="adapterMapperFactory"><see cref="IAdapterMapperFactory"/> for creating mappers for adapters.</param>
         public Adapters(
             ITypes types,
             IServiceProvider serviceProvider,
-            IEventStore eventStore,
-            IEventTypes eventTypes,
-            JsonProjectionSerializer projectionSerializer,
-            ILoggerFactory loggerFactory)
+            IAdapterProjectionFactory adapterProjectionFactory,
+            IAdapterMapperFactory adapterMapperFactory)
         {
             _types = types;
             _serviceProvider = serviceProvider;
-            _eventStore = eventStore;
-            _eventTypes = eventTypes;
-            _projectionSerializer = projectionSerializer;
-            _loggerFactory = loggerFactory;
+            _adapterProjectionFactory = adapterProjectionFactory;
+            _adapterMapperFactory = adapterMapperFactory;
             PopulateAdapters();
         }
 
@@ -63,7 +47,7 @@ namespace Aksio.Integration
             ThrowIfMissingAdapterForModelAndExternalModel<TModel, TExternalModel>();
             if (AdaptersByKey<TModel, TExternalModel>.Projection is null)
             {
-                CreateProjectionFor<TModel, TExternalModel>();
+                AdaptersByKey<TModel, TExternalModel>.Projection = _adapterProjectionFactory.CreateFor(AdaptersByKey<TModel, TExternalModel>.Adapter!);
             }
             return AdaptersByKey<TModel, TExternalModel>.Projection!;
         }
@@ -74,62 +58,23 @@ namespace Aksio.Integration
             ThrowIfMissingAdapterForModelAndExternalModel<TModel, TExternalModel>();
             if (AdaptersByKey<TModel, TExternalModel>.Mapper is null)
             {
-                CreateMapperFor<TModel, TExternalModel>();
+                AdaptersByKey<TModel, TExternalModel>.Mapper = _adapterMapperFactory.CreateFor(AdaptersByKey<TModel, TExternalModel>.Adapter!);
             }
             return AdaptersByKey<TModel, TExternalModel>.Mapper!;
         }
 
         void PopulateAdapters()
         {
+            var adaptersByKeyType = typeof(AdaptersByKey<,>);
+
             foreach (var adapterType in _types.FindMultiple(typeof(IAdapterFor<,>)))
             {
                 var adapterInterface = adapterType.GetInterface(typeof(IAdapterFor<,>).Name)!;
-                var adaptersByKey = typeof(AdaptersByKey<,>).MakeGenericType(adapterInterface.GenericTypeArguments);
+                var adaptersByKey = adaptersByKeyType.MakeGenericType(adapterInterface.GenericTypeArguments);
                 var adapterProperty = adaptersByKey.GetField(nameof(AdaptersByKey<object, object>.Adapter), BindingFlags.Public | BindingFlags.Static)!;
                 var adapter = _serviceProvider.GetService(adapterType);
                 adapterProperty.SetValue(null, adapter);
             }
-        }
-
-        void CreateProjectionFor<TModel, TExternalModel>()
-        {
-            ThrowIfMissingAdapterForModelAndExternalModel<TModel, TExternalModel>();
-
-            var projectionBuilder = new ProjectionBuilderFor<TModel>(Guid.Empty, _eventTypes);
-            AdaptersByKey<TModel, TExternalModel>.Adapter!.DefineModel(projectionBuilder);
-            var definition = projectionBuilder.Build();
-
-            var converters = new JsonConverter[]
-            {
-                new ConceptAsJsonConverter(),
-                new ConceptAsDictionaryJsonConverter()
-            };
-
-            var json = JsonConvert.SerializeObject(definition, converters);
-            var parsed = _projectionSerializer.Deserialize(json);
-            var projection = _projectionSerializer.CreateFrom(parsed);
-
-            AdaptersByKey<TModel, TExternalModel>.Projection = new AdapterProjectionFor<TModel>(projection, _eventStore.GetStream(EventStreamId.EventLog), _loggerFactory);
-        }
-
-        void CreateMapperFor<TModel, TExternalModel>()
-        {
-            ThrowIfMissingAdapterForModelAndExternalModel<TModel, TExternalModel>();
-            var configuration = new MapperConfiguration(cfg =>
-            {
-                cfg.ShouldUseConstructor = ci =>
-                {
-                    var parameters = ci.GetParameters();
-                    return !ci.IsPrivate && !(parameters.Length == 1 && parameters[0].ParameterType.Equals(ci.DeclaringType));
-                };
-                cfg.ShouldMapMethod = mi => false;
-                cfg.ShouldMapField = fi => !fi.IsPrivate;
-                cfg.AllowNullDestinationValues = true;
-                var mapping = cfg.CreateMap<TExternalModel, TModel>();
-                mapping = mapping.DisableCtorValidation();
-                AdaptersByKey<TModel, TExternalModel>.Adapter!.DefineImportMapping(mapping!);
-            });
-            AdaptersByKey<TModel, TExternalModel>.Mapper = configuration.CreateMapper();
         }
 
         void ThrowIfMissingAdapterForModelAndExternalModel<TModel, TExternalModel>()
