+++ Source/Integration/IAdapterProjectionFactory.cs
@@ -1,0 +1,17 @@
+namespace Aksio.Integration
+{
+    /// <summary>
+    /// Defines a factory for creating <see cref="IAdapterProjectionFor{TModel}"/>.
+    /// </summary>
+    public interface IAdapterProjectionFactory
+    {
+        /// <summary>
+        /// Create a <see cref="IAdapterProjectionFor{TModel}"/> from an <see cref="IAdapterFor{TModel, TExternalModel}"/>.
+        /// </summary>
+        /// <param name="adapter"><see cref="IAdapterFor{TModel, TExternalModel}"/> to create for.</param>
+        /// <typeparam name="TModel">Type of model.</typeparam>
+        /// <typeparam name="TExternalModel">Type of external model.</typeparam>
+        /// <returns>A new instance of <see cref="IAdapterProjectionFor{TModel}"/>.</returns>
+        IAdapterProjectionFor<TModel> CreateFor<TModel, TExternalModel>(IAdapterFor<TModel, TExternalModel> adapter);
+    }
+}
